Moved alias to aliases. Removed tjm from package.json
