+++ packages/teraslice-cli/cmds/aliases/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-exports.command = 'alias <command>';
+exports.command = 'aliases <command>';
 exports.desc = 'commands to add and remove cluster aliases';
 exports.exclude = 'lib';
 exports.builder = function builder(yargs) {
