+++ packages/teraslice-cli/cmds/aliases/lib/index.js
