+++ packages/teraslice-cli/cmds/aliases/list.js
@@ -1,6 +1,3 @@
-// update existing cluster
-// create new
-
 'use strict';
 'use console';
 
@@ -9,20 +6,18 @@ const reply = require('../lib/reply')();
 const config = require('../lib/config');
 const cli = require('../lib/cli');
 
-exports.command = 'remove';
+exports.command = 'list';
 exports.desc = 'List the clusters defined in the config file.\n';
 exports.builder = (yargs) => {
-    cli().args('alias', 'remove', yargs);
-    yargs
-        .demandCommand(1);
-    yargs.example('earl alias remove cluster1');
+    cli().args('alias', 'list', yargs);
+    yargs.example('earl aliases list cluster1');
 };
 
 exports.handler = (argv, _testFunctions) => {
     const cliConfig = _.clone(argv);
-    config(cliConfig, 'alias:remove').returnConfigData(false, false);
-    const libAlias = _testFunctions || require('./lib')(cliConfig);
+    config(cliConfig, 'aliases:list').returnConfigData(false);
+    const libAliases = _testFunctions || require('./lib')(cliConfig);
 
-    return libAlias.remove()
+    return libAliases.list()
         .catch(err => reply.fatal(err.message));
 };
