+++ packages/teraslice-cli/cmds/aliases/remove.js
@@ -9,20 +9,20 @@ const reply = require('../lib/reply')();
 const config = require('../lib/config');
 const cli = require('../lib/cli');
 
-exports.command = 'add';
-exports.desc = 'Add an alias to the clusters defined in the config file.\n';
+exports.command = 'remove';
+exports.desc = 'List the clusters defined in the config file.\n';
 exports.builder = (yargs) => {
-    cli().args('alias', 'list', yargs);
+    cli().args('aliases', 'remove', yargs);
     yargs
         .demandCommand(1);
-    yargs.example('earl alias add cluster1 -c http://cluster1.net:80');
+    yargs.example('earl aliases remove cluster1');
 };
 
 exports.handler = (argv, _testFunctions) => {
     const cliConfig = _.clone(argv);
-    config(cliConfig, 'alias:add').returnConfigData(false, false);
-    const libAlias = _testFunctions || require('./lib')(cliConfig);
+    config(cliConfig, 'aliases:remove').returnConfigData(false, false);
+    const libAliases = _testFunctions || require('./lib')(cliConfig);
 
-    return libAlias.add()
+    return libAliases.remove()
         .catch(err => reply.fatal(err.message));
 };
