+++ packages/teraslice-cli/package.json
@@ -4,7 +4,6 @@
   "description": "Teraslice command line interface",
   "bin": {
     "teraslice-cli": "./index.js",
-    "tjm": "./index.js",
     "earl": "./index.js"
   },
   "main": "index.js",
